# docs/swagger.yaml
basePath: /api/1
definitions:
  Error:
    properties:
      error:
        type: string
    type: object
  List:
    properties:
      count:
        type: integer
      word:
        type: string
    type: object
  RelatedLinks:
    properties:
      sid:
        type: string
      tid:
        type: string
    type: object
  RelatedNodes:
    properties:
      _size:
        type: number
      id:
        type: string
      name:
        type: string
    type: object
  RelatedTweet:
    properties:
      createdAt:
        type: string
      text:
        type: string
    type: object
  RelatedTweets:
    properties:
      code:
        type: integer
      message:
        items:
          $ref: '#/definitions/RelatedTweet'
        type: array
    type: object
  Tag:
    properties:
      percent:
        type: number
      tag:
        type: string
    type: object
  TagCount:
    properties:
      count:
        type: integer
      tag:
        type: string
    type: object
  WordCloud:
    properties:
      words:
        items:
          type: object
        type: array
    type: object
  WordCount:
    properties:
      article:
        items:
          type: integer
        type: array
      community:
        items:
          type: integer
        type: array
      date:
        items:
          type: string
        type: array
      sns:
        items:
          type: integer
        type: array
      total:
        items:
          type: integer
        type: array
    type: object
  WordsResponse:
    properties:
      links:
        items:
          $ref: '#/definitions/RelatedLinks'
        type: array
      nodes:
        items:
          $ref: '#/definitions/RelatedNodes'
        type: array
      rank:
        additionalProperties:
          type: number
        type: object
    type: object
host: api.whatissuenow.com
info:
  contact:
    email: 6199@outlook.kr
    name: Jintae, kim
    url: http://whatissuenow.com
  description: WhatIssueNow Service API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: WhatIssueNow API
  version: "1.1"
paths:
  /list:
    get:
      consumes:
      - application/json
      description: 현재시간 기준 3시간 전까지의 상위 이슈 단어 30개를 반환한다
      parameters:
      - description: Word list count by count
        in: query
        name: count
        type: integer
      - description: Word Filtering by f
        in: query
        name: f
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/List'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      summary: 이슈 단어목록 API
      tags:
      - WhatIssueNow
  /related/list/{word}:
    get:
      consumes:
      - application/json
      description: 특정 단어의 관련된 최근 트윗 목록 100개를 반환한다
      parameters:
      - description: Word
        in: path
        name: word
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RelatedTweets'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      summary: 특정 단어의 최근 트윗 목록 API
      tags:
      - WhatIssueNow
  /related/w/{word}:
    get:
      consumes:
      - application/json
      description: 특정 단어와 관련된 다른 단어들을 반환한다
      parameters:
      - description: Word
        in: path
        name: word
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/WordsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      summary: 특정 단어와 연관된 단어 목록 API
      tags:
      - WhatIssueNow
  /stopwords:
    get:
      consumes:
      - application/json
      description: ElasticSearch 집계시에 제외되는 불용어 목록을 반환한다
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
      summary: 불용어 단어목록 API
      tags:
      - StopWords
  /tag/count:
    get:
      consumes:
      - application/json
      description: 현재시간 기준 3시간 전까지의 태그별 단어 수를 반환한다
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/TagCount'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      summary: 태그별 단어 수 API
      tags:
      - WhatIssueNow
  /tag/w/{word}:
    get:
      consumes:
      - application/json
      description: 특정 단어의 발생지(태그) 점유율을 반환한다
      parameters:
      - description: Word
        in: path
        name: word
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Tag'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      summary: 단어별 태그 점유율 API
      tags:
      - WhatIssueNow
  /word/count:
    get:
      consumes:
      - application/json
      description: 현재시간 기준 7일전까지의 수집한 단어의 수를 일별로 반환한다
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/WordCount'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      summary: 일주일간 수집한 단어의 수 API
      tags:
      - WhatIssueNow
  /wordcloud:
    get:
      consumes:
      - application/json
      description: 현재시간 기준 3시간 전까지의 상위 이슈 단어 60개를 반환한다
      parameters:
      - description: Word list count by count
        in: query
        name: count
        type: integer
      - description: Word Filtering by f
        in: query
        name: f
        type: string
      - description: 워드클라우드 단어 개수(최대 100개)
        in: query
        name: count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/WordCloud'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
      summary: 워드 클라우드를 위한 이슈 단어목록 API
      tags:
      - WhatIssueNow
schemes:
- http
- https
swagger: "2.0"